Cleanup: Moved Methods to Helper.js
